requirements.txt
boto3
torch
torchvision
pillow
numpy
pyspark
